refactor(menu): improve stories snippet parsing
