+++ src/components/Menu/__stories__/Borderless.stories.tsx
@@ -1,46 +1,47 @@
+import { ComponentStory } from '@storybook/react'
 import { fireEvent, screen } from '@storybook/testing-library'
 import Menu from '..'
-import { Template } from './Template'
+import { DefaultDisclosure } from './Template.stories'
 
-export const Borderless = Template.bind({})
-
-Borderless.parameters = {
-  docs: {
-    storyDescription: 'Property `borderless` removes border of the menu item.',
-  },
-}
-
-Borderless.args = {
-  children: [
-    <Menu.Item borderless>default</Menu.Item>,
+export const Borderless: ComponentStory<typeof Menu> = () => (
+  <Menu disclosure={DefaultDisclosure}>
+    <Menu.Item borderless>default</Menu.Item>
     <Menu.Item variant="danger" borderless>
       Danger
-    </Menu.Item>,
+    </Menu.Item>
     <Menu.Item variant="nav" borderless>
       Nav
-    </Menu.Item>,
-    <Menu.Item borderless>default</Menu.Item>,
+    </Menu.Item>
+    <Menu.Item borderless>default</Menu.Item>
     <Menu.Item
       href="/?path=/docs/components-navigation-menu--borderless"
       borderless
     >
       Link
-    </Menu.Item>,
+    </Menu.Item>
+
     <Menu.Item
       variant="danger"
       href="/?path=/docs/components-navigation-menu--borderless"
       borderless
     >
       Link Danger
-    </Menu.Item>,
+    </Menu.Item>
+
     <Menu.Item
       variant="nav"
       href="/?path=/docs/components-navigation-menu--borderless"
       borderless
     >
       Link Nav
-    </Menu.Item>,
-  ],
+    </Menu.Item>
+  </Menu>
+)
+
+Borderless.parameters = {
+  docs: {
+    storyDescription: 'Property `borderless` removes border of the menu item.',
+  },
 }
 
 Borderless.play = () => {
