+++ src/components/Menu/__stories__/WithModal.stories.tsx
@@ -1,22 +1,11 @@
+import { ComponentStory } from '@storybook/react'
 import { fireEvent, screen } from '@storybook/testing-library'
 import Menu from '..'
 import Modal from '../../Modal'
-import { Template } from './Template'
+import { DefaultDisclosure } from './Template.stories'
 
-export const WithModal = Template.bind({})
-
-WithModal.parameters = {
-  docs: {
-    storyDescription: 'This show how to use a modal on MenuItem.',
-  },
-}
-
-WithModal.args = {
-  children: [
-    <Menu.Item>Menu Item</Menu.Item>,
-    <Menu.Item href="/?path=/docs/components-navigation-menu--modal">
-      Menu Item Link
-    </Menu.Item>,
+export const WithModal: ComponentStory<typeof Menu> = () => {
+  const NestedModal = () => (
     <Modal
       animated
       animation="scaleUp"
@@ -28,8 +17,24 @@ WithModal.args = {
       <div style={{ padding: 32 }}>
         Content should be present in center of the modal
       </div>
-    </Modal>,
-  ],
+    </Modal>
+  )
+
+  return (
+    <Menu disclosure={DefaultDisclosure}>
+      <Menu.Item>Menu Item</Menu.Item>
+      <Menu.Item href="/?path=/docs/components-navigation-menu--modal">
+        Menu Item Link
+      </Menu.Item>
+      <NestedModal />
+    </Menu>
+  )
+}
+
+WithModal.parameters = {
+  docs: {
+    storyDescription: 'This show how to use a modal on MenuItem.',
+  },
 }
 
 WithModal.play = async () => {
